kdirstat: Correct output syntax for FIFO
There needs to be some whitspace after the FIFO tag, otherwise k4dirstat
(silently!) stops parsing at this entry.

This was wrong in the initial commit to add FIFO support to kdirstat
4ce858d7c593 ("kdirstat: Handled named pipes (fifos)"), manual testing at the
time apparently only verified that the file was generated, not that it was
valid, sorry.

I updated the test yarns to also check for FIFO, however this is only run
"ASSUMING extended attributes are allowed for users", which does not appear to
be the case on my system. Also the testshould probably test for all sorts of
files, not just regular and FIFO.

Signed-off-by: Ian Campbell <ijc@hellion.org.uk>