Include the root dir into kdirstat output
Root directory is now excluded from kdirstat output done by obnam.
That solution was probably a workaround to a k4dirstat bug, however,
this broke displaying any --root="/" backups and since we now have
qdirstat (and k4dirstat became deprecated) which fixed the bug, it is
time to reintroduce the root dir into the kdirstat output. Discussion
with Ian about this can be read in the thread "Kdirstat merged to
Qdirstat".